:sparkles: :: [#289] LectureDetailSettingFeature / 강의 학과 PopupView -> BottomSheet로 변경
