+++ App/Sources/Feature/LectureDetailSettingFeature/Sources/Component/BottomSheet/DepartmentBottomSheet.swift
@@ -1,29 +1,29 @@
 import SwiftUI
 
-struct DivisionBottomSheet: View {
-    var selectedDivision: String
+struct DepartmentBottomSheet: View {
+    var selectedDepartment: String
     @Binding var keyword: String
     @State var isSelected: Bool = false
-    let divisionList: [String]
-    let divisionDidSelect: (String) -> Void
+    let departmentList: [String]
+    let departmentSelect: (String) -> Void
 
     var body: some View {
-        VStack {
+        VStack(spacing: 8) {
             SearchTextField(
                 text: $keyword,
-                "강의 구분 검색"
+                "강의 학과 검색"
             )
-
+            
             ScrollView(showsIndicators: false) {
-                if divisionList.isEmpty {
+                if departmentList.isEmpty {
                     HStack {
                         VStack(alignment: .leading) {
                             BitgouelText(
                                 text: keyword,
                                 font: .text2
                             )
-
-                            Text("새 계열 추가하기")
+                            
+                            Text("새 학과 추가하기")
                                 .bitgouelFont(.caption, color: .greyscale(.g4))
                         }
 
@@ -35,7 +35,7 @@ struct DivisionBottomSheet: View {
                                 set: { selected in
                                     if selected {
                                         isSelected = selected
-                                        divisionDidSelect(keyword)
+                                        departmentSelect(keyword)
                                     } else {
                                         isSelected = selected
                                     }
@@ -45,16 +45,14 @@ struct DivisionBottomSheet: View {
                     }
                     .padding(.vertical, 24)
                 }
-
-                LazyVStack {
-                    ForEach(divisionList, id: \.self) { division in
-                        divisionListRow(
-                            division: division,
+                
+                LazyVStack(alignment: .leading) {
+                    ForEach(departmentList, id: \.self) { department in
+                        departmentRow(
+                            department: department,
                             isSelected: Binding(
-                                get: { division == selectedDivision },
-                                set: { _ in
-                                    print(division)
-                                }
+                                get: { department == selectedDepartment},
+                                set: { _ in departmentSelect(department) }
                             )
                         )
                     }
@@ -65,22 +63,21 @@ struct DivisionBottomSheet: View {
     }
 
     @ViewBuilder
-    func divisionListRow(
-        division: String,
+    func departmentRow(
+        department: String,
         isSelected: Binding<Bool>
     ) -> some View {
         HStack {
-            VStack(alignment: .leading) {
-                BitgouelText(
-                    text: division,
-                    font: .text2
-                )
-            }
+            BitgouelText(
+                text: department,
+                font: .text1
+            )
 
             Spacer()
 
             BitgouelRadioButton(isSelected: isSelected)
         }
+        .id(department)
         .padding(.vertical, 24)
     }
 }
